requirements_dev.txt
flask
flask-cors
pyserial
pytest
pytest-cov
flake8
autopep8
pylint
pyflakes
pytest-mock
mock
